requirements.txt
appdirs==1.4.3
Django==1.11.1
django-admin-view-permission==0.8
django-crispy-forms==1.6.1
django-registration-redux==1.6
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
